feat: refactor event and tickettype modules with fragments
introduce `fragments.actions` and `fragments.policies` to the
`gits.storefront.event` module, and `fragments.calculations` to
the `gits.storefront.tickettype` module. this refactor improves
modularity and organization of actions and policies, enhancing
code maintainability. additionally, update the `show` live view
to load and display the total ticket types for events.